Inserting a Genus
Ok, back to the database-world. Let's be really lazy and skip creating a form - there's a whole series on forms later, and, I'd just hate to spoil the fun.

Instead, insert some hardcoded data. How? Simple: start with {dollar_sign}genus = new Genus(), put some data on that object, and tell Doctrine to save it.

Doctrine wants you to stop thinking about queries, and instead think about objects.

Right now... name is the only real field we have. And it's a private property, so we can't actually put data on it. To make this mutable - to use a really fancy term that means 'changeable' - go to the bottom of the class and open the 'Code'->'Generate' menu. Select 'Getters and Setters' - we'll need the getter function later

Great! Now use {dollar_sign}genus->setName() and call it Octopus with a random ending to make things more fun!