Debugging with the Web Debug Toolbar
Go back to the controller and hack in some HTML markup into the response.

Try it again! Ah, there you are fancy web debug toolbar. There are actually three database queries. Interesting. Click the icon to enter the profiler. Ah, there's the insert query, hiding inside a transaction.

And by the way, how sweet is this for debugging? You can see a formatted query, a runnable version, or run EXPLAIN on a slow query.

Running SQL Queries in the Terminal

I still can't believe it's working - things never work on the first try! To triple-check it, head to the terminal. To run a raw SQL query, use:

./bin/console doctrine:query:sql 'SELECT * FROM genus'
There they are. So inserting objects with Doctrine... pretty darn easy.