We're close to making the genus page dynamic - but it has a few fields that we don't have yet - like sub family, number of known species, and fun fact.
Let's add these... and be as lazy as possible when we do it! In Genus, create the properties first: private {dollar_sign}subFamily, private {dollar_sign}speciesCount and private {dollar_sign}funFact

Now, just do the same thing we did before: bring up the 'Code'->'Generate' menu - or command+N on a Mac - select 'ORM Annotation' and choose all the fields to generate the Column annotations above each. These tries to guess the right field type - but it's not always right. speciesCount should clearly be an integer - set it to that. For a full-list of all the built-in Doctrine types, check their docs. The most important are string, integer, text, datetime and float.