Updating the Table Schema
Now, can we tell Doctrine to somehow alter the genus table and add these columns? Absolutely. In fact, it's one of the most incredible features of Doctrine.

In the terminal, run:

./bin/console doctrine:schema:update --dump-sql
Look familiar? That's the same command we ran before. But look: it actually says, 'ALTER TABLE genus' add sub_family, species_count, and fun_fact. This is amazing. It's able to look at the database, look at the entity, and calculating the difference between them.

So if we were to run this with --force, it would run that query and life would be good. Should we do it? No! Wait, hold on!

It would work. But imagine our app was already deployed and working on production with the first version of the genus table. When you deploy the new code, you'll need to run this command on production to update your table.

And that'll work 99% of the time. But sometimes, the query might not be perfect. For example, what if I rename a property? Well, this command might drop the existing column and add a new one. All the data from the old column would be gone! The point is: running doctrine:schema:update is just too dangerous on production.

But, we're going to replace it with something just as good.