Copy the new statement from the docs and paste that into the AppKernel class.
Beautiful!

We already know that the main job of a bundle is to give us new services. But this bundle primarily gives us something different: a new set of console commands. Run bin/console with no arguments:

./bin/console
Hiding in the middle is a whole group starting with doctrine:migrations. These are our new best friend.