Making Columns nullable
Here's the deal: Doctrine configures all columns to be required in the database by default. If you do want a column to be 'nullable', find the column and add nullable=true.

Of course, just because we made this change doesn't mean that our table was automatically updated behind the scenes.