Open up the tr. Bring this to life with a td that prints {{ genus.name }} and another that prints {{ genus.speciesCount }}. And hey, we're getting autocompletion, that's kind of nice.
The Magic Twig '.' Notation

Refresh! Easy - it looks exactly how we want it. But wait a second... something cool just happened in the background. We printed genus.name... but name is a private property - so we should not be able to access it directly. How is this working?
81 lines  src/AppBundle/Entity/Genus.php
... lines 1 - 10
class Genus {
... lines 13 - 22
  private {dollar_sign}name;
... lines 24 - 39
  public function getName(){
    return {dollar_sign}this->name;
  }
... lines 44 - 79
}

This is Twig to the rescue! Behind the scenes, Twig noticed that name was private and called getName() instead. And it does the same thing with genus.speciesCount:

 81 lines  src/AppBundle/Entity/Genus.php
... lines 1 - 10
class Genus {
... lines 13 - 32
  private {dollar_sign}speciesCount;
... lines 34 - 59
  public function getSpeciesCount(){
    return {dollar_sign}this->speciesCount;
  }
... lines 64 - 79
}
Twig is smart enough to figure out how to access the data - and this lets us keep the template simple.