Twig is smart enough to figure out how to access the data - and this lets us keep the template simple.
With that in mind, I have a challenge! Add a third column to the table called 'Last Updated'

This won't work yet, but what I I want to be able to say is {{ genus.updatedAt }}. If this existed and returned a DateTime object, we could pipe it through the built-in Twig date filter to format it.
But this won't work - there is not an updatedAt property. We'll add one later, but we're stuck right now.