We have a list page! Heck, we have a show page. Let's link them together.
First, the poor show route is nameless. Give it a name - and a new reason to live - with name='genus_show'

That sounds good.

In the list template, and the a tag and use the path() function to point this to the genus_show route. Remember - this route has a {genusName} wildcard, so we must pass a value for that here. Add a set of curly-braces to make an array... But this is getting a little long: so break onto multiple lines. Much better. Finish with genusName: genus.name. And make sure the text is still genus.name.

Cool! Refresh. Oooh, pretty links. Click the first one. The name is 'Octopus66', but the fun fact and other stuff is still hardcoded. It's time to grow up and finally make this dynamic!