Querying for One Genus
In the controller, get rid of . We need to query for a Genus that matches the . First, fetch the entity manager with {dollar_sign}em = {dollar_sign}this->getDoctrine()->getManager().

Then, {dollar_sign}genus = {dollar_sign}em->getRepository() with the AppBundle:Genus shortcut. Ok now, is there a method that can help us? Ah, how about findOneBy(). This works by passing it an array of things to find by - in our case 'name' => {dollar_sign}genusName.

Oh, and comment out the caching for now - it's temporarily going to get in the way.

Get outta here caching!

Finally, since we have a Genus object, we can simplify the render() call and only pass it.
Open up show.html.twig: we just changed the variables passed into this template, so we've got work to do. First, use genus.name and then genus.name again. Remove the hardcoded sadness and replace it with genus.subFamily, genus.speciesCount and genus.funFact. Oh, and remove the raw filter - we're temporarily not rendering this through markdown. Put it on the todo list.
There's one more spot down in the JavaScript - change this to genus.name.
Okay team, let's give it a try. Refresh. Looks awesome! The known species is the number it should be, there is no fun fact, and the JavaScript is still working.