But what would happen if somebody went to a genus name that did not exist - like FOOBARFAKENAMEILOVEOCTOPUS? Woh! We get a bad error. This is coming from Twig:
Impossible to access an attribute ('name') on a null variable
because on line 3, genus is null - it's not a Genus object:
40 lines  app/Resources/views/genus/show.html.twig
... lines 1 - 2
{% block title %}Genus {{ genus.name }}{% endblock %}
In the prod environment, this would be a 500 page. We do not want that - we want the user to see a nice 404 page, ideally with something really funny on it.