Handling 404's
Back in the controller, the findOneBy() method will either return one Genus object or null. If it does not return an object, throw {dollar_sign}this->createNotFoundException('No genus found')

Oh, and that message will only be shown to developers - not to end-users.

Head back, refresh, and this is a 404. In the prod environment, the user will see a 404 template that you need to setup. I won't cover how to customize the template here - it's pretty easy - just make sure it's really clever, and send me a screenshot. Do it!