Now things are about to get fun. A few minutes ago, we installed a library called nelmio/alice(https://github.com/nelmio/alice) - search for that and find their GitHub page.
In a nutshell, this library lets us add fixtures data via YAML files. It has an expressive syntax and it ships with a bunch of built-in functions for generating random data. Actually, it uses yet another library behind the scenes called Faker(https://github.com/fzaninotto/Faker) to do that. It's the PHP circle of life!